refactor: change dht values to float

+++ src/variables/variables.cpp
@@ -24,11 +24,11 @@ extern const uint8_t buzzerPin = 22;
 
 extern const uint8_t dhtPin = 14;
 
-int16_t temperature;
-int16_t humidity;
+float temperature;
+float humidity;
 
-int16_t currentTemperature;
-int16_t currentHumidity;
+float currentTemperature;
+float currentHumidity;
 
 // O intervalo em milissegundos em que o usuário pode apertar o botão | The interval in milliseconds that the user can press the touch button
 unsigned long touchInterval;
